requirements.txt
Django>=2.0,<3.0
django-cors-headers==3.0.1
djangorestframework==3.9.4
psycopg2
Celery
django-extensions